school-timetable/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==0.21.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
pydantic==1.10.7
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
